move `g3w-ol/controls/control.js` into `g3w-ol/controls/interactioncontrol.js`
